Fixes the LIDAR depth perception for Lincoln MKZ.
The previous code would normalize the point cloud, and find the point
closest to our 2D point transformed to 3D. However, this would fail
because the closest point may not be near the same (x, y) coordinate.
This fix utilizes just the (x, y) coordinate to figure out the closest
point in the point cloud, and utilizes its z-value instead of
normalizing.

Also, we don't have to take the unreal transform if we are driving the
real vehicle.